split out vector_of_kll python class from kll_wrapper, address gcc9 warnings on fallthrough in murmur hash
