+++ python/src/datasketches.cpp
@@ -27,6 +27,7 @@ void init_fi(py::module& m);
 void init_cpc(py::module& m);
 void init_theta(py::module& m);
 void init_vo(py::module& m);
+void init_vector_of_kll(py::module& m);
 
 PYBIND11_MODULE(datasketches, m) {
   init_hll(m);
@@ -35,4 +36,5 @@ PYBIND11_MODULE(datasketches, m) {
   init_cpc(m);
   init_theta(m);
   init_vo(m);
+  init_vector_of_kll(m);
 }
